store: allow configurable scan properties in scan-range
Large range scans may happen in the record layer by either
scanning explicit ranges or execute queries for which the
plan determines a scan should be performed. While this is not
something one should optimize for, it is legitimate to want
to perform large scans on occasion.

In this case, it may be ok to allow relaxing consistency
guarantees. To achieve that, we introduce a low-level
`continuation-traversing-reducer` function, used by
`long-range-reducer` and `long-query-reducer`.

A lot of help from @arnaudgeiser, especially around fixing
prefix range building.